Moved Fail View creation to FailViewFactory so that it can be used in other Use Cases
